github_student_inviter/requirements.txt
PyGithub
pandas
python-dotenv
requests

